// frontend/src/app/core/api/generated/api/warehouses.service.ts
export interface FindAllRequestParams {
    page?: number;
    size?: number;
}